+++ Libft/ft_memcpy.c
@@ -6,17 +6,17 @@
 /*   By: agesp <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 10:41:48 by agesp             #+#    #+#             */
-/*   Updated: 2018/11/08 16:16:42 by agesp            ###   ########.fr       */
+/*   Updated: 2018/11/09 11:39:34 by agesp            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, size_t n)
+void				*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int i;
-	unsigned char *savedst;
-	unsigned char *savesrc;
+	size_t			i;
+	unsigned char	*savedst;
+	unsigned char	*savesrc;
 
 	i = 0;
 	savedst = (unsigned char*)dst;
@@ -31,5 +31,3 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 	dst = savedst;
 	return (dst);
 }
-
-
