+++ Libft/ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: agesp <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 15:06:59 by agesp             #+#    #+#             */
-/*   Updated: 2018/11/07 16:56:48 by agesp            ###   ########.fr       */
+/*   Updated: 2018/11/09 11:41:57 by agesp            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void		*ft_memset(void *b, int c, size_t len)
 {
-	int i;
-	char *save;
+	size_t	i;
+	char	*save;
 
 	i = 0;
 	save = (char*)b;
@@ -27,5 +27,3 @@ void	*ft_memset(void *b, int c, size_t len)
 	b = save;
 	return (b);
 }
-
-
